move keras and fasttext to models module
