+++ docs/source/cleanlab/experimental/index.rst
@@ -11,8 +11,6 @@ experimental
    :show-inheritance:
 
 .. toctree::
-    keras
-    fasttext
     mnist_pytorch
     coteaching
     cifar_cnn
