+++ docs/source/cleanlab/models/index.rst
@@ -1,0 +1,15 @@
+models
+======
+
+.. warning::
+    Methods in this ``models`` module are not guaranteed to be stable between different ``cleanlab`` versions.
+
+.. automodule:: cleanlab.models
+   :autosummary:
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+.. toctree::
+    keras
+    fasttext
